refactor ui inspector drawer
